+++ designsystem/src/main/res/drawable/img_school.png
