+++ designsystem/src/main/res/drawable/img_student.png
