+++ designsystem/src/main/res/drawable/img_teacher.png
