+++ feature-main/workspace/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    alias(libs.plugins.seugi.android.feature)
+}
+
+android {
+    namespace = "com.seugi.workspace"
+}
+
+dependencies {
+    implementation(projects.designsystem)
+    implementation(projects.data.member)
+    implementation(projects.data.workspace)
+    implementation(projects.common)
+    implementation(projects.ui)
+}
