+++ feature-main/workspace/src/main/java/com/seugi/workspace/feature/WaitingJoinNavigation.kt
@@ -1,10 +1,9 @@
-package com.seugi.join.navigation
+package com.seugi.workspace.feature
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.seugi.join.feature.WaitingJoinScreen
 
 const val WAITING_JOIN = "waitingJoin"
 
@@ -15,6 +14,6 @@ fun NavController.navigateToWaitingJoin(navOptions: NavOptions? = null) = naviga
 
 fun NavGraphBuilder.waitingJoin(popBackStack: () -> Unit) {
     composable(route = WAITING_JOIN) {
-        WaitingJoinScreen(popBackStack = popBackStack)
+        com.seugi.workspace.feature.WaitingJoinScreen(popBackStack = popBackStack)
     }
 }
