+++ feature-main/workspace/src/main/java/com/seugi/workspace/feature/schoolcode/SchoolCodeScreen.kt
@@ -1,4 +1,4 @@
-package com.seugi.join.feature.schoolcode
+package com.seugi.workspace.feature.schoolcode
 
 import android.widget.Toast
 import androidx.compose.foundation.clickable
@@ -31,7 +31,7 @@ import com.seugi.designsystem.component.SeugiFullWidthButton
 import com.seugi.designsystem.component.SeugiTopBar
 import com.seugi.designsystem.component.textfield.SeugiCodeTextField
 import com.seugi.designsystem.theme.SeugiTheme
-import com.seugi.join.feature.schoolcode.model.SchoolCodeSideEffect
+import com.seugi.workspace.feature.schoolcode.model.SchoolCodeSideEffect
 import kotlinx.coroutines.flow.collectLatest
 
 @OptIn(ExperimentalMaterial3Api::class)
