+++ feature-main/workspace/src/main/java/com/seugi/workspace/feature/selectingcode/SelectingCodeViewModel.kt
@@ -1,4 +1,4 @@
-package com.seugi.join.feature.selectingcode
+package com.seugi.workspace.feature.selectingcode
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,7 +6,7 @@ import com.seugi.common.model.Result
 import com.seugi.common.utiles.DispatcherType
 import com.seugi.common.utiles.SeugiDispatcher
 import com.seugi.data.workspace.WorkspaceRepository
-import com.seugi.join.feature.selectingcode.model.SelectingCodeSideEffect
+import com.seugi.workspace.feature.selectingcode.model.SelectingCodeSideEffect
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
