+++ feature-main/workspace/src/main/java/com/seugi/workspace/navigation/SchoolCodeNavigation.kt
@@ -1,10 +1,10 @@
-package com.seugi.join.navigation
+package com.seugi.workspace.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.seugi.join.feature.schoolcode.SchoolScreen
+import com.seugi.workspace.feature.schoolcode.SchoolScreen
 
 const val SCHOOL_CODE = "schoolCode"
